fix: resolve layout primitive guard, name validation, and post ad price ux
