+++ apps/web/src/components/user/Login.tsx
@@ -30,7 +30,10 @@ import { Card, CardContent, CardHeader, CardTitle } from "../ui/card";
 // Validation Schemas
 const loginSchema = z.object({
   mobile: z.string().regex(/^\d{10}$/, "Please enter a valid 10-digit mobile number"),
-  name: z.string().optional(),
+  name: z.string()
+    .regex(/^[a-zA-Z\s'.-]*$/, "Name can only contain letters, spaces, dots, hyphens, and apostrophes")
+    .refine(val => !val || val.trim().length >= 2, { message: "Name must be at least 2 characters" })
+    .optional(),
   otp: z.string().optional(),
 });
